Refactoring: Move all report renderers to `grafana_wtf.report`
