+++ CHANGES.rst
@@ -9,6 +9,7 @@ in progress
 - Caching: Increase default cache TTL to five minutes again
 - Caching: Optionally configure TTL using environment variable ``CACHE_TTL``
 - History: Stop ``grafana-wtf log <UID>`` acquiring *all* dashboards
+- Refactoring: Move all report renderers to ``grafana_wtf.report``
 
 2023-03-05 0.14.1
 =================
